2897 client - Extract buildPropertyMentionsContent into propertyMentionDom
Not part of #2897, but it travels with it: upstream, this extraction and the
function autocomplete were folded into a single commit, and the callers that grew
out of it (the copilot apply path, and the controlled tool inputs in #2445 that
follow) import the helper by name. Splitting it out keeps the autocomplete commit
to the autocomplete.

The value-to-editor-HTML conversion was a 45-line closure inside the editor's
getContent callback, unreachable from anywhere else. It is the forward direction
of replaceMentionNodesInHtmlWithVariables, which already lives in propertyMentionDom,
so it belongs beside it. Behavior is unchanged; getContent now delegates.
